rebased chanes to master
Pulled update from master to rebase the changes onto the new master. The changes are the implementation of meta file sets for units that they get created from (in preparation for upgrates).

In the process some changes (added unitkeys to spawnactions, some name changes, added the databases to the scene) got lost and need to be readded at a later point